--- backend/smart-logistics/src/main/java/com/douzon/smartlogistics/domain/member/interceptor/ExceptionAdviceHandler.java
@@ -1,27 +1,0 @@
-package com.douzon.smartlogistics.domain.member.interceptor;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.nio.charset.StandardCharsets;
-
-@RestControllerAdvice
-public class ExceptionAdviceHandler {
-
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ExceptionHandler(AuthException.class)
-    public ResponseEntity<String> authException(AuthException e) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(new MediaType(MediaType.TEXT_PLAIN, StandardCharsets.UTF_8));
-
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .headers(headers)
-                .body("인증오류: " + e.getLocalizedMessage());
-    }
-
-}
