+++ backend/smart-logistics/src/main/java/com/douzon/smartlogistics/global/common/exception/auth/AuthException.java
@@ -1,4 +1,4 @@
-package com.douzon.smartlogistics.domain.member.interceptor;
+package com.douzon.smartlogistics.global.common.exception.auth;
 
 public class AuthException extends RuntimeException {
     public AuthException(String msg) {
